Converse/nlu/requirements.txt
scikit-learn==0.24.1
typer==0.3.0
pytorch-pretrained-BERT==0.6.2
transformers==4.5.1 spacy==3.0.6
ftfy==4.4.3
duckling==1.8.0
grpcio==1.37.1
grpcio-tools
protobuf==3.15.8
Jpype1==0.7.5
